+++ torchtext/experimental/datasets/__init__.py
@@ -1,3 +1,8 @@
+from .language_modeling import LanguageModelingDataset, WikiText2, WikiText103, PennTreebank  # NOQA
 from .text_classification import IMDB
 
-__all__ = ['IMDB']
+__all__ = ['LanguageModelingDataset',
+           'WikiText2',
+           'WikiText103',
+           'PennTreebank',
+           'IMDB']
